# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  review.Review:
    properties:
      customer_id:
        type: string
      description:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      review:
        type: number
    type: object
  review.ReviewRequest:
    properties:
      customer_id:
        type: string
      description:
        type: string
      post_id:
        type: integer
      review:
        type: number
    type: object
host: localhost:8070
info:
  contact:
    email: asliddinvstalim@gmail.com
    name: Asliddin
    url: https://t.me/asliddindeh
  description: This is reivew api
  termsOfService: not much usefull
  title: Review api
  version: "1.0"
paths:
...
  /review:
    post:
      consumes:
      - application/json
      description: this func create review
      parameters:
      - description: Review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/review.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.Review'
      security:
      - BearerAuth: []
      summary: create review
      tags:
      - review
  /review/{id}:
    get:
      consumes:
      - application/json
      description: this func get review by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.Review'
      security:
      - BearerAuth: []
      summary: get review by id
      tags:
      - review
  /review/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this func delete review
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
      security:
      - BearerAuth: []
      summary: delete review
      tags:
      - review
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"